backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.27.3
httpx==0.28.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
groq==1.0.0
huggingface-hub==0.19.4
redis==5.0.1
cryptography==41.0.7
PyJWT==2.8.0
bcrypt==3.2.0
gunicorn==21.2.0